build: Put generated sources into build folder
Sometimes (especially after an update), the xpp submodule would contain
outdated header files which could interfere with the compilation. These
header files then needed to be manually deleted. Now the header files
are in the build folder and so clearing the build folder will also clear
the header files.

The proto header files are now located at:

    build/lib/xpp/generated-sources/include/xpp/proto

Because the proto files are now no longer in the same place as the other
header files, all `#include` directives for including proto headers no
longer work and need to be adjusted.