Use python that provides xcbgen for generating
Currently we search with both python 2 and 3 for xcbgen. In some cases
(e.g. Ubuntu 17.10) this doesn't work because the python2 xcbgen syntax
that is no longer valid in python3. In other cases this does work
because the python2 version ships with updated syntax.

We now use whatever python executable was used to find xcbgen for
generating all the header files. This way we don't run into this syntax
error issue

Fixes polybar/polybar#2159